requirements.txt
cffi==1.15.1
colorama==0.4.6
crayons==0.4.0
numpy==1.23.4
opencv-python==4.6.0.66
piexif==1.1.3
Pillow==9.3.0
pycparser==2.21
PyNaCl==1.5.0
stegano==0.10.2
